[All] Handling additional cases of potential null pointer exceptions

+++ illaeasyquest/src/illarion/easyquest/gui/IntegerParameter.java
@@ -53,6 +53,6 @@ public class IntegerParameter extends JFormattedTextField implements Parameter {
 
     @Nonnull
     public Object getParameter() {
-        return (Long) getValue();
+        return getValue();
     }
 }
